Move instance handle build into reader
