+++ dds/src/implementation/dds_impl/user_defined_subscriber.rs
@@ -1,5 +1,4 @@
 use crate::implementation::rtps::endpoint::RtpsEndpoint;
-use crate::implementation::rtps::instance_handle_builder::InstanceHandleBuilder;
 use crate::implementation::rtps::messages::submessages::{DataSubmessage, HeartbeatSubmessage};
 use crate::implementation::rtps::reader::RtpsReader;
 use crate::implementation::rtps::transport::TransportWrite;
@@ -117,7 +116,7 @@ impl DdsShared<UserDefinedSubscriber> {
                 false => TopicKind::NoKey,
             };
 
-            let rtps_reader = RtpsStatefulReader::new(RtpsReader::new(
+            let rtps_reader = RtpsStatefulReader::new(RtpsReader::new::<Foo>(
                 RtpsEndpoint::new(
                     guid,
                     topic_kind,
@@ -128,7 +127,6 @@ impl DdsShared<UserDefinedSubscriber> {
                 DURATION_ZERO,
                 false,
                 qos,
-                InstanceHandleBuilder::new::<Foo>(),
             ));
 
             let data_reader_shared = UserDefinedDataReader::new(
